requirements.txt
anytree==2.12.1
bezier==2024.6.20
fa2_modified==0.3.9
matplotlib==3.10.1
mdtraj==1.9.9
msmhelper==1.1.1
networkx==3.1
numba==0.59.1
numpy==2.2.4
pandas==2.2.3
prettypyplot==0.11.0
pymol==0.1.0
PyYAML==6.0.2
scikit_learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
setuptools==70.0.0
tqdm==4.66.2
